requirements.txt
streamlit
transformers
torch
nltk
scikit-learn
wordcloud
python-docx
PyPDF2
sentencepiece
rouge-score
python-docx
PyPDF2
